# Don't preserve newlines in attributes