`Haml::Helpers.preserve` is very slow. I don't think it's necessary for attributes because Slim, Faml and Hamlit don't preserve them and work fine. 

The original change was done in 2008 https://github.com/haml/haml/commit/e8075e7feac72291a203b137b4f8c2a13457e843 and no reason is found to keep that change. Let's remove a legacy feature.

## Benchmark
With Ruby 2.4.0 and  [k0kubun/haml_bench/templates/slim_bench.haml](https://github.com/k0kubun/haml_bench/blob/97a7b9970684b975c48e1403ee4fe329bf1151fb/templates/slim_bench.haml),

### before
```
$ bundle exec ruby bench.rb
Rendering: /home/k0kubun/src/github.com/k0kubun/haml_bench/templates/slim_bench.haml
Calculating -------------------------------------
          haml 4.0.7     3.576k i/100ms
   haml 5.0.0.beta.2     5.450k i/100ms
-------------------------------------------------
          haml 4.0.7     37.559k (± 3.6%) i/s -    189.528k
   haml 5.0.0.beta.2     55.625k (± 4.5%) i/s -    277.950k

Comparison:
   haml 5.0.0.beta.2:    55625.3 i/s
          haml 4.0.7:    37558.8 i/s - 1.48x slower
```

### after
```
$ bundle exec ruby bench.rb
Rendering: /home/k0kubun/src/github.com/k0kubun/haml_bench/templates/slim_bench.haml
Calculating -------------------------------------
          haml 4.0.7     3.608k i/100ms
   haml 5.0.0.beta.2     6.140k i/100ms
-------------------------------------------------
          haml 4.0.7     37.460k (± 3.5%) i/s -    187.616k
   haml 5.0.0.beta.2     64.248k (± 4.5%) i/s -    325.420k

Comparison:
   haml 5.0.0.beta.2:    64247.5 i/s
          haml 4.0.7:    37459.9 i/s - 1.72x slower
```

@haml/committers WDYT?