@k0kubun 2017-02-09
I finished attribute compiler implementation using this. Merging for now. Please give me your opinions later.